// CLIENT/src/apis/stock/response/editShopWarehouse.dto.ts
interface EditWarehouseDTO {
  address?: string | '';

  city?: string | '';

  code?: string | '';

  createdAt?: number;

  createdBy?: string | '';

  id?: string | '';

  modifiedAt?: number;

  modifiedBy?: string | '';

  name?: string | '';

  phoneNumber?: string | '';

  postalCode?: string | '';

  prefectureCode?: string | '';

  shopId?: string | '';

  status?: number;
}
export default EditWarehouseDTO;